requirements.txt
Flask==0.9
python-swiftclient>=1.8.0
python-novaclient>=2.26.0
